feat: student dashboard
