+++ lms/public/css/style.css
@@ -326,7 +326,7 @@ input[type=checkbox] {
 	border-radius: var(--border-radius-md);
 	position: relative;
 	border: 1px solid var(--gray-300);
-	box-shadow: var(--card-shadow);
+	box-shadow: var(--shadow-inset);
 }
 
 .course-card {
@@ -1650,10 +1650,6 @@ li {
 	color: var(--gray-600);
 }
 
-.discussions-parent .empty-state {
-	background-color: var(--gray-200);
-}
-
 .job-cards-parent {
 	display: grid;
 	grid-template-columns: repeat(auto-fill, minmax(500px, 1fr));
@@ -2204,4 +2200,77 @@ select {
 .rows .grid-row .grid-footer-toolbar,
 .grid-form-heading {
 	cursor: none;
+}
+
+.lms-page-style .discussions-section-title {
+	font-size: var(--text-lg);
+}
+
+.class-dashboard .progress {
+	width: 150px;
+	height: 150px;
+	background: none;
+	position: relative;
+}
+
+.class-dashboard .progress::after {
+	content: "";
+	width: 100%;
+	height: 100%;
+	border-radius: 50%;
+	border: 6px solid #eee;
+	position: absolute;
+	top: 0;
+	left: 0;
+}
+
+.class-dashboard .progress>span {
+	width: 50%;
+	height: 100%;
+	overflow: hidden;
+	position: absolute;
+	top: 0;
+	z-index: 1;
+}
+
+.class-dashboard .progress .progress-left {
+	left: 0;
+}
+
+.class-dashboard .progress .progress-bar {
+	width: 100%;
+	height: 100%;
+	background: none;
+	border-width: 6px;
+	border-style: solid;
+	position: absolute;
+	top: 0;
+}
+
+.class-dashboard .progress .progress-left .progress-bar {
+	left: 100%;
+	border-top-right-radius: 80px;
+	border-bottom-right-radius: 80px;
+	border-left: 0;
+	-webkit-transform-origin: center left;
+	transform-origin: center left;
+}
+
+.class-dashboard .progress .progress-right {
+	right: 0;
+}
+
+.class-dashboard .progress .progress-right .progress-bar {
+	left: -100%;
+	border-top-left-radius: 80px;
+	border-bottom-left-radius: 80px;
+	border-right: 0;
+	-webkit-transform-origin: center right;
+	transform-origin: center right;
+}
+
+.class-dashboard .progress .progress-value {
+	position: absolute;
+	top: 0;
+	left: 0;
 }
