+++ lms/www/classes/progress.html
@@ -64,112 +64,13 @@
 
 {% macro UpcomingEvals(upcoming_evals) %}
 <div class="mb-8">
-	<div class="bold-heading mb-2">
-		{{ _("Upcoming Evaluations") }}
-	</div>
-	{% if upcoming_evals | length %}
-	<article class="cards-parent">
-		{% for eval in upcoming_evals %}
-		<div class="common-card-style column-card">
-			<div class="flex align-center justify-between">
-				<div class="bold-heading">
-					{{ eval.course_title }}
-				</div>
-				{% if eval.google_meet_link %}
-				<a class="btn btn-default btn-sm pull-right" href="{{ eval.google_meet_link }}">
-					{{ _("Join") }}
-				</a>
-				{% endif %}
-			</div>
-
-			<div class="vertically-center">
-				<svg class="icon icon-sm mr-1">
-					<use href="#icon-calendar"></use>
-				</svg>
-				<span>
-					{{ frappe.utils.format_date(eval.date, "medium") }}&nbsp;-&nbsp;
-				</span>
-				<span>
-					{{ frappe.utils.format_time(eval.start_time, "hh:mm a") }}
-				</span>
-			</div>
-			<div>
-				<span class="field-label">
-					{{ _("Evaluator") }}:
-				</span>
-				<span>
-					{{ eval.evaluator_name }}
-				</span>
-			</div>
-		</div>
-		{% endfor %}
-	</article>
-	{% else %}
-	<p class="text-muted"> {{ _("No Upcoming Evaluations") }} </p>
-	{% endif %}
+	{% include "lms/templates/upcoming_evals.html" %}
 </div>
 {% endmacro %}
 
 {% macro Assessments(class_info, student) %}
 <div class="mb-8">
-	<div class="bold-heading mb-2">
-		{{ _("Assessments") }}
-	</div>
-	{% if assessments | length %}
-	<article class="form-grid">
-		<div class="grid-heading-row">
-			<div class="grid-row">
-				<div class="data-row row">
-					<div class="col grid-static-col">
-						{{ _("Assessment") }}
-					</div>
-					<div class="col grid-static-col col-xs-2">
-						{{ _("Type") }}
-					</div>
-					<div class="col grid-static-col col-xs-2">
-						{{ _("Status/Score") }}
-					</div>
-				</div>
-			</div>
-		</div>
-		{% for assessment in assessments %}
-		{% set has_access = is_moderator and assessment.submission or frappe.session.user == student.name %}
-		<div class="grid-row">
-			<div class="data-row row">
-				<a class="col grid-static-col {% if has_access %} clickable {% endif %}" {% if has_access %} href="{{ assessment.url }}" {% endif %}>
-					{{ assessment.title }}
-				</a>
-				<div class="col grid-static-col col-xs-2">
-					{{ (assessment.assessment_type).split("LMS ")[1] }}
-				</div>
-
-				<div class="col grid-static-col col-xs-2 mb-2">
-					{% if assessment.submission %}
-						{% if assessment.assessment_type == "LMS Assignment" %}
-							{% set status = assessment.submission.status %}
-							{% set color = "green" if status == "Pass" else "red" if status == "Fail" else "orange"  %}
-							<div class="indicator-pill {{ color }}">
-								{{ status }}
-							</div>
-						{% else %}
-							<div>
-								{{ assessment.submission.score }}
-							</div>
-						{% endif %}
-					{% else %}
-						<div class="indicator-pill red">
-							{{ _("Not Attempted") }}
-						</div>
-					{% endif %}
-				</div>
-
-			</div>
-		</div>
-		{% endfor %}
-	</article>
-	{% else %}
-	<p class="text-muted"> {{ _("No Assessments") }} </p>
-	{% endif %}
+	{% include "lms/templates/assessments.html" %}
 </div>
 
 {% endmacro %}
