+++ lms/www/classes/progress.py
@@ -1,5 +1,9 @@
 import frappe
-from lms.lms.utils import has_course_moderator_role, has_course_evaluator_role
+from lms.lms.utils import (
+	has_course_moderator_role,
+	has_course_evaluator_role,
+	get_upcoming_evals,
+)
 from frappe import _
 from lms.www.utils import get_assessments
 
@@ -34,20 +38,4 @@ def get_context(context):
 	)
 
 	context.assessments = get_assessments(class_name, context.student.name)
-
-	upcoming_evals = frappe.get_all(
-		"LMS Certificate Request",
-		{
-			"member": context.student.name,
-			"course": ["in", context.courses],
-			"date": [">=", frappe.utils.nowdate()],
-		},
-		["date", "start_time", "course", "evaluator", "google_meet_link"],
-		order_by="date",
-	)
-
-	for evals in upcoming_evals:
-		evals.course_title = frappe.db.get_value("LMS Course", evals.course, "title")
-		evals.evaluator_name = frappe.db.get_value("User", evals.evaluator, "full_name")
-
-	context.upcoming_evals = upcoming_evals
+	context.upcoming_evals = get_upcoming_evals(context.student.name, context.courses)
